+++ app/models/task.rb
@@ -1,7 +1,6 @@
-class Section < ApplicationRecord
+class Task < ApplicationRecord
   self.implicit_order_column = "created_at"
-  belongs_to :journey
-  has_many :tasks
+  belongs_to :section
 
   validates :title, :contentful_id, presence: true
 end